Refactor: `CommentActions` folder structure
