+++ src/modules/core/components/Comment/Actions/CommentActionsPopover.css
