+++ src/modules/core/components/Comment/Actions/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CommentActions';
